requirements.txt
streamlit
langchain
langchain-openai
python-dotenv
openai
langserve
fastapi
uvicorn
langchain-groq
langchain-community
sse_starlette